+++ programmable_video/android/src/main/kotlin/twilio/flutter/twilio_programmable_video/VideoCapturerHandler.kt
@@ -8,10 +8,10 @@ import android.hardware.camera2.CameraMetadata
 import android.hardware.camera2.CaptureRequest
 import com.twilio.video.Camera2Capturer
 import com.twilio.video.CameraCapturer
-import com.twilio.video.VideoCapturer
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import java.lang.Exception
+import tvi.webrtc.VideoCapturer
 
 class VideoCapturerHandler {
     companion object {
@@ -25,36 +25,20 @@ class VideoCapturerHandler {
         }
 
         @JvmStatic
-        fun switchCamera(call: MethodCall, result: MethodChannel.Result) {
-            if (TwilioProgrammableVideoPlugin.cameraCapturer != null && TwilioProgrammableVideoPlugin.cameraCapturer is Camera2Capturer) {
-                return switchCamera2Capturer(call, result)
-            } else if (TwilioProgrammableVideoPlugin.cameraCapturer != null && TwilioProgrammableVideoPlugin.cameraCapturer is CameraCapturer) {
-                return switchCameraCapturer(call, result)
-            }
-            return result.error("NOT_FOUND", "No CameraCapturer has been initialized yet, try connecting first.", null)
-        }
-
-        @JvmStatic
-        fun hasTorch(result: MethodChannel.Result) {
-            var hasTorch = hasTorch()
-            TwilioProgrammableVideoPlugin.debug("VideoCapturerHandler::hasTorch => check: $hasTorch")
-
-            result.success(hasTorch)
-        }
-
-        @JvmStatic
         fun setTorch(call: MethodCall, result: MethodChannel.Result) {
             TwilioProgrammableVideoPlugin.debug("VideoCapturerHandler.setTorch => called")
             val enableTorch = call.argument<Boolean>("enable")
-                ?: return result.error("MISSING_PARAMS", "The parameter 'enable' was not given", null)
+                    ?: return result.error("MISSING_PARAMS", "The parameter 'enable' was not given", null)
 
-            if (hasTorch() == true) {
-                if (TwilioProgrammableVideoPlugin.cameraCapturer == null) return result.error("FAILED", "Could not setTorch to enabled: $enableTorch, cameraCapturer is not defined", null)
-                else if (TwilioProgrammableVideoPlugin.cameraCapturer is Camera2Capturer) return setTorchCamera2Capturer(enableTorch, result)
-                else if (TwilioProgrammableVideoPlugin.cameraCapturer is CameraCapturer) return setTorchCameraCapturer(enableTorch, result)
-                else return result.error("FAILED", "Method `setTorch` not supported for ${TwilioProgrammableVideoPlugin.cameraCapturer.javaClass}", null)
+            return if (hasTorch()) {
+                when (TwilioProgrammableVideoPlugin.cameraCapturer) {
+                    null -> result.error("FAILED", "Could not setTorch to enabled: $enableTorch, cameraCapturer is not defined", null)
+                    is Camera2Capturer -> setTorchCamera2Capturer(enableTorch, result)
+                    is CameraCapturer -> setTorchCameraCapturer(enableTorch, result)
+                    else -> result.error("FAILED", "Method `setTorch` not supported for ${TwilioProgrammableVideoPlugin.cameraCapturer?.javaClass}", null)
+                }
             } else {
-                return result.error("FAILED", "Current camera does not have a flash", null)
+                result.error("FAILED", "Current camera does not have a flash", null)
             }
         }
 
@@ -63,29 +47,33 @@ class VideoCapturerHandler {
                 override fun onError(errorCode: Int) {
                     TwilioProgrammableVideoPlugin.handler.post {
                         TwilioProgrammableVideoPlugin.debug("CameraCapturer.onError => code: $errorCode")
-                        TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("cameraError", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer)), Exception(errorCode.toString()))
+                        TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("cameraError", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer!!)), Exception(errorCode.toString()))
                     }
                 }
 
                 override fun onFirstFrameAvailable() {
                     TwilioProgrammableVideoPlugin.handler.post {
                         TwilioProgrammableVideoPlugin.debug("CameraCapturer.onFirstFrameAvailable")
-                        TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("firstFrameAvailable", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer)), null)
+                        TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("firstFrameAvailable", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer!!)), null)
                     }
                 }
 
-                override fun onCameraSwitched() {
+                override fun onCameraSwitched(newCameraId: String) {
                     TwilioProgrammableVideoPlugin.handler.post {
-                        val cameraSource = (TwilioProgrammableVideoPlugin.cameraCapturer as CameraCapturer).cameraSource
-                        TwilioProgrammableVideoPlugin.debug("CameraCapturer.onCameraSwitched => newCameraSource: $cameraSource")
-                        TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("cameraSwitched", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer)), null)
+                        TwilioProgrammableVideoPlugin.debug("CameraCapturer.onCameraSwitched => newCameraId: $newCameraId")
+                        TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("cameraSwitched", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer!!, newCameraId)), null)
                     }
                 }
             }
-            val videoCapturer: VideoCapturer = when (videoCapturerMap["cameraSource"] as String) {
-                "BACK_CAMERA" -> CameraCapturer(TwilioProgrammableVideoPlugin.pluginHandler.applicationContext, CameraCapturer.CameraSource.BACK_CAMERA, listener)
-                else -> CameraCapturer(TwilioProgrammableVideoPlugin.pluginHandler.applicationContext, CameraCapturer.CameraSource.FRONT_CAMERA, listener)
-            }
+
+            val source = videoCapturerMap["source"] as Map<String, Any?>
+            val cameraId = source["cameraId"] as String?
+
+            // Check type because we may want to add support for ScreenCapturer
+            val videoCapturer: VideoCapturer = if (TwilioProgrammableVideoPlugin.cameraEnumerator.deviceNames.contains(cameraId))
+                CameraCapturer(TwilioProgrammableVideoPlugin.pluginHandler.applicationContext, cameraId!!, listener)
+            else
+                return result.error("MISSING_CAMERA", "No camera found for $cameraId.", null)
 
             if (videoCapturer is CameraCapturer) {
                 TwilioProgrammableVideoPlugin.cameraCapturer = videoCapturer
@@ -94,29 +82,29 @@ class VideoCapturerHandler {
 
         @JvmStatic
         private fun initializeCamera2Capturer(videoCapturerMap: Map<*, *>, result: MethodChannel.Result) {
-            // Check type because we may want to add support for ScreenCapturer
-            val cameraId = when (videoCapturerMap["cameraSource"] as String) {
-                "BACK_CAMERA" -> getCameraId(CameraMetadata.LENS_FACING_BACK)
-                else -> getCameraId(CameraMetadata.LENS_FACING_FRONT)
-            } ?: return result.error("MISSING_CAMERA", "No camera found for ${videoCapturerMap["cameraSource"]}.", null)
+            val source = videoCapturerMap["source"] as Map<String, Any?>
+            val cameraId = source["cameraId"] as String?
+            if (!TwilioProgrammableVideoPlugin.cameraEnumerator.deviceNames.contains(cameraId))
+                return result.error("MISSING_CAMERA", "No camera found for $cameraId.", null)
 
+            // Check type because we may want to add support for ScreenCapturer
             val videoCapturer: VideoCapturer = Camera2Capturer(
                     TwilioProgrammableVideoPlugin.pluginHandler.applicationContext,
-                    cameraId,
+                    cameraId!!,
                     object : Camera2Capturer.Listener {
                         override fun onError(camera2CapturerException: Camera2Capturer.Exception) {
                             TwilioProgrammableVideoPlugin.debug("Camera2Capturer.onError => $camera2CapturerException")
-                            TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("cameraError", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer)), camera2CapturerException)
+                            TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("cameraError", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer!!)), camera2CapturerException)
                         }
 
                         override fun onFirstFrameAvailable() {
                             TwilioProgrammableVideoPlugin.debug("Camera2Capturer.onFirstFrameAvailable")
-                            TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("firstFrameAvailable", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer)), null)
+                            TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("firstFrameAvailable", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer!!)), null)
                         }
 
                         override fun onCameraSwitched(newCameraId: String) {
                             TwilioProgrammableVideoPlugin.debug("Camera2Capturer.onCameraSwitched => newCameraId: $newCameraId")
-                            TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("cameraSwitched", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer)), null)
+                            TwilioProgrammableVideoPlugin.pluginHandler.sendCameraEvent("cameraSwitched", mapOf("capturer" to videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer!!)), null)
                         }
                     }
             )
@@ -125,45 +113,8 @@ class VideoCapturerHandler {
             }
         }
 
-        private fun switchCameraCapturer(call: MethodCall, result: MethodChannel.Result) {
-            val capturer = TwilioProgrammableVideoPlugin.cameraCapturer as CameraCapturer
-            val source = if (capturer.cameraSource == CameraCapturer.CameraSource.FRONT_CAMERA) {
-                CameraCapturer.CameraSource.BACK_CAMERA
-            } else {
-                CameraCapturer.CameraSource.FRONT_CAMERA
-            }
-            capturer.switchCamera()
-
-            return result.success(videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer, source))
-        }
-
-        private fun switchCamera2Capturer(call: MethodCall, result: MethodChannel.Result) {
-            val newCameraId: String?
-            val newCameraSource: CameraCapturer.CameraSource
-            val capturer = TwilioProgrammableVideoPlugin.cameraCapturer as Camera2Capturer
-            if (getCameraDirection(capturer.cameraId) == CameraMetadata.LENS_FACING_FRONT) {
-                newCameraId = getCameraId(CameraMetadata.LENS_FACING_BACK)
-                newCameraSource = CameraCapturer.CameraSource.BACK_CAMERA
-            } else {
-                newCameraId = getCameraId(CameraMetadata.LENS_FACING_FRONT)
-                newCameraSource = CameraCapturer.CameraSource.FRONT_CAMERA
-            }
-
-            if (newCameraId != null) {
-                capturer.switchCamera(newCameraId)
-                return result.success(videoCapturerToMap(TwilioProgrammableVideoPlugin.cameraCapturer, newCameraSource))
-            } else {
-                return result.error("MISSING_CAMERA", "Could not find another camera to switch to", null)
-            }
-        }
-
         private fun getCameraManager(): CameraManager {
             return TwilioProgrammableVideoPlugin.pluginHandler.applicationContext.getSystemService(Context.CAMERA_SERVICE) as CameraManager
-        }
-
-        private fun getCameraId(cameraDirection: Int): String? {
-            val cameraManager: CameraManager = getCameraManager()
-            return cameraManager.cameraIdList.firstOrNull { cameraId -> cameraManager.getCameraCharacteristics(cameraId)[CameraCharacteristics.LENS_FACING] == cameraDirection }
         }
 
         private fun getCameraDirection(cameraId: String): Int? {
@@ -172,23 +123,15 @@ class VideoCapturerHandler {
             return cameraManager.getCameraCharacteristics(cameraId)[CameraCharacteristics.LENS_FACING]
         }
 
-        private fun getCameraDirectionAsString(direction: Int?): String {
-            when (direction) {
-                CameraMetadata.LENS_FACING_FRONT -> return "FRONT_CAMERA"
-                CameraMetadata.LENS_FACING_BACK -> return "BACK_CAMERA"
-                else -> return "UNKNOWN"
-            }
-        }
-
         private fun cameraIdCorrespondsToActiveCamera(capturer: CameraCapturer, id: String): Boolean {
-            var cameraInfo = Camera.CameraInfo()
+            val cameraInfo = Camera.CameraInfo()
             Camera.getCameraInfo(id.toInt(), cameraInfo)
-            return if (capturer?.cameraSource == null)
-                false
-            else if (capturer.cameraSource == CameraCapturer.CameraSource.FRONT_CAMERA)
-                cameraInfo.facing == Camera.CameraInfo.CAMERA_FACING_FRONT
-            else
-                cameraInfo.facing == Camera.CameraInfo.CAMERA_FACING_BACK
+
+            return when (getCameraDirection(capturer.cameraId)) {
+                CameraMetadata.LENS_FACING_FRONT -> cameraInfo.facing == Camera.CameraInfo.CAMERA_FACING_FRONT
+                CameraMetadata.LENS_FACING_BACK -> cameraInfo.facing == Camera.CameraInfo.CAMERA_FACING_BACK
+                else -> false
+            }
         }
 
         private fun hasTorchCameraCapturer(): Boolean {
@@ -214,10 +157,12 @@ class VideoCapturerHandler {
         }
 
         private fun hasTorch(): Boolean {
-            if (TwilioProgrammableVideoPlugin.cameraCapturer == null) return false
-            else if (TwilioProgrammableVideoPlugin.cameraCapturer is Camera2Capturer) return hasTorchCamera2Capturer()
-            else if (TwilioProgrammableVideoPlugin.cameraCapturer is CameraCapturer) return hasTorchCameraCapturer()
-            else return false
+            return when (TwilioProgrammableVideoPlugin.cameraCapturer) {
+                null -> false
+                is Camera2Capturer -> hasTorchCamera2Capturer()
+                is CameraCapturer -> hasTorchCameraCapturer()
+                else -> false
+            }
         }
 
         private fun setTorchCameraCapturer(enableTorch: Boolean, result: MethodChannel.Result) {
@@ -229,10 +174,10 @@ class VideoCapturerHandler {
                 }
             }
 
-            if (scheduled) {
-                return result.success(null)
+            return if (scheduled) {
+                result.success(null)
             } else {
-                return result.error("FAILED", "Failed to schedule updateCaptureRequest", null)
+                result.error("FAILED", "Failed to schedule updateCaptureRequest", null)
             }
         }
 
@@ -248,31 +193,45 @@ class VideoCapturerHandler {
                 it.set(CaptureRequest.FLASH_MODE, flashMode)
             }
 
-            if (scheduled) {
-                return result.success(null)
+            return if (scheduled) {
+                result.success(null)
             } else {
-                return result.error("FAILED", "Failed to schedule updateCaptureRequest", null)
+                result.error("FAILED", "Failed to schedule updateCaptureRequest", null)
             }
         }
 
-        fun videoCapturerToMap(videoCapturer: VideoCapturer, cameraSource: CameraCapturer.CameraSource? = null): Map<String, Any> {
+        @JvmStatic
+        fun cameraIdToMap(cameraId: String): Map<String, Any> {
+            val cameraManager: CameraManager = getCameraManager()
+            val hasTorch = cameraManager.getCameraCharacteristics(cameraId)[CameraCharacteristics.FLASH_INFO_AVAILABLE]
+                    ?: false
+
+            return mapOf(
+                    "isFrontFacing" to TwilioProgrammableVideoPlugin.cameraEnumerator.isFrontFacing(cameraId),
+                    "isBackFacing" to TwilioProgrammableVideoPlugin.cameraEnumerator.isBackFacing(cameraId),
+                    "hasTorch" to hasTorch,
+                    "cameraId" to cameraId
+            )
+        }
+
+        fun videoCapturerToMap(videoCapturer: VideoCapturer, cameraId: String? = null): Map<String, Any> {
             if (videoCapturer is Camera2Capturer) {
-                var source = getCameraDirectionAsString(getCameraDirection(videoCapturer.cameraId))
-                if (cameraSource != null) {
-                    source = cameraSource.toString()
+                var id = videoCapturer.cameraId
+                if (cameraId != null) {
+                    id = cameraId.toString()
                 }
                 return mapOf(
                         "type" to "CameraCapturer",
-                        "cameraSource" to source
+                        "source" to cameraIdToMap(id)
                 )
             } else if (videoCapturer is CameraCapturer) {
-                var source = videoCapturer.cameraSource.toString()
-                if (cameraSource != null) {
-                    source = cameraSource.toString()
+                var id = videoCapturer.cameraId
+                if (cameraId != null) {
+                    id = cameraId
                 }
                 return mapOf(
                         "type" to "CameraCapturer",
-                        "cameraSource" to source
+                        "source" to cameraIdToMap(id)
                 )
             }
             return mapOf(