+++ programmable_video/lib/src/events/room_events.dart
@@ -5,7 +5,7 @@ class RoomConnectFailureEvent {
   final Room room;
 
   /// An exception describing why connect failed.
-  final TwilioException exception;
+  final TwilioException? exception;
 
   RoomConnectFailureEvent(
     this.room,
@@ -20,7 +20,7 @@ class RoomReconnectingEvent {
   /// An error explaining why the [LocalParticipant] is reconnecting to a room.
   /// Errors are limited to [TwilioException.signalingConnectionDisconnectedException]
   /// and [TwilioException.mediaConnectionErrorException].
-  final TwilioException exception;
+  final TwilioException? exception;
 
   RoomReconnectingEvent(
     this.room,