+++ programmable_video_platform_interface/test/models/tracks/local_video_track_model_test.dart
@@ -1,12 +1,11 @@
-import 'package:enum_to_string/enum_to_string.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:twilio_programmable_video_platform_interface/src/enums/camera_source.dart';
+import 'package:twilio_programmable_video_platform_interface/src/camera_source.dart';
 import 'package:twilio_programmable_video_platform_interface/src/models/model_exports.dart';
 
 void main() {
   final name = 'name';
   final enabled = true;
-  final cameraSource = CameraSource.FRONT_CAMERA;
+  final source = CameraSource('FRONT_CAMERA', true, false, true);
   final type = 'CameraCapturer';
 
   group('.fromEventChannelMap()', () {
@@ -14,12 +13,15 @@ void main() {
       final map = {
         'name': name,
         'enabled': enabled,
-        'videoCapturer': {'cameraSource': EnumToString.convertToString(cameraSource), 'type': type}
+        'videoCapturer': {'source': source.toMap(), 'type': type}
       };
       final model = LocalVideoTrackModel.fromEventChannelMap(map);
       expect(model.name, name);
       expect(model.enabled, enabled);
-      expect(model.cameraCapturer.source, cameraSource);
+      expect(model.cameraCapturer.source?.cameraId, source.cameraId);
+      expect(model.cameraCapturer.source?.isFrontFacing, source.isFrontFacing);
+      expect(model.cameraCapturer.source?.isBackFacing, source.isBackFacing);
+      expect(model.cameraCapturer.source?.hasTorch, source.hasTorch);
       expect(model.cameraCapturer.type, type);
     });
 
@@ -34,18 +36,18 @@ void main() {
       final model = LocalVideoTrackModel(
         name: name,
         enabled: enabled,
-        cameraCapturer: CameraCapturerModel(cameraSource, type),
+        cameraCapturer: CameraCapturerModel(source, type),
       );
       expect(
         model.toString(),
-        '{ name: $name, enabled: $enabled, cameraCapturer: { source: $cameraSource, type: $type, isScreencast: false } }',
+        '{ name: $name, enabled: $enabled, cameraCapturer: { source: $source, type: $type, isScreencast: false } }',
       );
     });
   });
 
   group('.toMap()', () {
     test('should return correct Map', () {
-      final cameraCapturer = CameraCapturerModel(cameraSource, type);
+      final cameraCapturer = CameraCapturerModel(source, type);
       final model = LocalVideoTrackModel(
         name: name,
         enabled: enabled,
@@ -55,7 +57,7 @@ void main() {
         'enable': enabled,
         'name': name,
         'videoCapturer': {
-          'cameraSource': EnumToString.convertToString(cameraCapturer.source),
+          'source': source.toMap(),
           'type': cameraCapturer.type,
         },
       });