+++ programmable_video_platform_interface/CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.6.0+1
+
+- Bumped minor versions of dependencies
+
+## 0.6.0
+
+- **BREAKING**: Replaced `CameraSource` enum with a class variant to represent a potential source for camera capturing.
+- **BREAKING**: Removed `hasTorch()` from the interface as this value is now accessible on the new `CameraSource` class.
+- **BREAKING**: `switchCamera()` now expects a `CameraSource` as an argument, it should switch to that source when called.
+- Added `getSources()` on the interface which should return all potential `CameraSource`s a device has.
+
+## 0.5.0
+
+- **BREAKING**: Made `exception` property of ConnectFailure, Reconnecting and Disconnected events nullable.
+
 ## 0.4.0+1
 
 - Resolve "'!' used on CameraSource enum which does not include 'null'"